refactor(pagination): move and split type in a separated module
