+++ packages/kit-headless/src/components/pagination/types/index.ts
@@ -1,0 +1,1 @@
+export * from './pagination-api';
