+++ packages/kit-headless/src/components/pagination/types/pagination-api.ts
@@ -1,0 +1,31 @@
+import { PropFunction } from '@builder.io/qwik';
+
+export interface PaginationCoreProps {
+  selectedPage: number;
+  totalPages: number;
+  onPageChange$: PropFunction<(page: number) => void>;
+}
+
+export type PaginationStyling = {
+  class?: string;
+  gap?: string;
+  defaultClass?: string;
+  selectedClass?: string;
+  dividerClass?: string;
+};
+
+export type PaginationConfig = {
+  siblingCount?: number;
+  boundaryCount?: number;
+  hidePrevButton?: boolean;
+  hideNextButton?: boolean;
+  disabled?: boolean;
+  customArrowTexts?: ArrowLabels;
+};
+
+type ArrowLabels = {
+  previous: string;
+  next: string;
+};
+
+export type PaginationProps = PaginationCoreProps & PaginationStyling & PaginationConfig;
